requirements.txt
openai==1.30.5
pillow==10.4.0
pygments==2.19.2
opencv-python==4.13.0.92
numpy==1.24.4
gtts==2.5.4
edge-tts==6.1.12
google-cloud-texttospeech==2.17.2
pyttsx3==2.99
reportlab==4.4.3
python-docx==1.1.2
python-pptx==1.0.2
requests==2.32.4
httpx==0.27.2
